// EscalafonFrontend/src/app/interfaces/legajo-ubicacion.ts
export interface LegajoUbicacion {
    /**
     * @type { number } - iLegUbiId
     */
    iLegUbiId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { string } - cLegUbiAmbiente
     */
    cLegUbiAmbiente?: string;

    /**
     * @type { string } - cLegUbiEstante
     */
    cLegUbiEstante?: string;

    /**
     * @type { string } - cLegUbiCaja
     */
    cLegUbiCaja?: string;

    /**
     * @type { number } - cLegUbiSeccion1Folios
     */
    cLegUbiSeccion1Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion2Folios
     */
    cLegUbiSeccion2Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion3Folios
     */
    cLegUbiSeccion3Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion4Folios
     */
    cLegUbiSeccion4Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion5Folios
     */
    cLegUbiSeccion5Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion6Folios
     */
    cLegUbiSeccion6Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion7Folios
     */
    cLegUbiSeccion7Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion8Folios
     */
    cLegUbiSeccion8Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion9Folios
     */
    cLegUbiSeccion9Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion10Folios
     */
    cLegUbiSeccion10Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion11Folios
     */
    cLegUbiSeccion11Folios?: number;

    /**
     * @type { number } - cLegUbiSeccion12Folios
     */
    cLegUbiSeccion12Folios?: number;

}